BTS: check for expect before trying install

+++ modules/rhizo_base/files/var/SysmoBTS/provision.sh
@@ -18,11 +18,14 @@ for bts in "${!BTS[@]}" ; do
   ssh $SSH_OPTS root@${BTS[$bts]} "date -s '$(date)'"
   ssh $SSH_OPTS root@${BTS[$bts]} "echo nameserver 1.1.1.1 > /etc/resolv.conf; echo nameserver 9.9.9.9 >> /etc/resolv.conf"
   scp $SSH_OPTS osmo-pcu.cfg root@${BTS[$bts]}:/etc/osmocom/osmo-pcu.cfg
-  ssh $SSH_OPTS root@${BTS[$bts]} "mkdir /tmp/ipk"
-  scp $SSH_OPTS *.ipk root@${BTS[$bts]}:/tmp/ipk/
-  ssh $SSH_OPTS root@${BTS[$bts]} "opkg install /tmp/ipk/*.ipk"
-  ssh $SSH_OPTS root@${BTS[$bts]} "rm -r /tmp/ipk"
+  ssh $SSH_OPTS root@${BTS[$bts]} "expect -v 2>/dev/null"
 
+  if [ "$?" != "0" ]; then
+    ssh $SSH_OPTS root@${BTS[$bts]} "mkdir /tmp/ipk"
+    scp $SSH_OPTS *.ipk root@${BTS[$bts]}:/tmp/ipk/
+    ssh $SSH_OPTS root@${BTS[$bts]} "opkg install /tmp/ipk/*.ipk"
+    ssh $SSH_OPTS root@${BTS[$bts]} "rm -r /tmp/ipk"
+  fi
 
   _modelNR=$(ssh $SSH_OPTS ${BTS[$bts]} sysmobts-util model-nr)
